refactor: ArgNavBar to own file
